# Conflicts:
#	ruoyi-system/src/main/java/com/ruoyi/system/service/impl/DownloadreportServiceImpl.java
#	ruoyi-system/src/main/resources/mapper/system/XianFactorAnalysisMapper.xml

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/dto/DisasterRainDTO.java
@@ -10,8 +10,8 @@ public class DisasterRainDTO {
     private Long disasterId;
     private String rainfall;
     private String duration;
-    private Long longitude;
-    private Long latitude;
+    private Double longitude;
+    private Double latitude;
     private String position;
     private String disasterName;
     private LocalDateTime occurrenceTime;
